Remove additional drawing orientation for Y axis
This one was responsible for flipping the Y axis numbers from leftwards
to rightwards when the slope of the Y axis on screen becomes negative.
IMO, this is slightly weird because as viewer you have to readjust the
orientation in which you read, but more importantly this is inconsistent
with how the X axis is drawn (which simply doesn't make this
distinction). I think both axes should be drawn the same way.