Add option 'showGrayBottom'
Previously, this depended on the value of showShadow, which IMO is a
bug, because it couples two independent decisions.

The option was mentioned as not working, because of (I'm guessing) two
problems: First, the top/bottom-side calculation bug addressed in the
previous commit. Second, because there are edge-cases (literally) where
where a surface is curved such that part of it is viewed from below and
another part from above. To properly address this scenario we would need
different drawing styles for the individual triangles of the surface
polygon.

However, this scenario occurs so rarely, and it even then looks good
enough to me. Furthermore, since this option is new and disabled by
default, I doubt we will make anyone unhappy by giving them this choice.

I made this new option visible in "Styling Surface", because I think it
fits there the most, and it should be visible at least one of the
examples.