Improve lighting calculation
The previous behaviour looked incomprehensible to me. Now it scales by
how much the surface is oriented toward the viewer, and makes the bottom
of the surface darker. This corresponds to diffuse lighting with light
shining in the view direction.

Alternatively, we could:

- make no distinction between bottom and top:
  v = Math.abs(crossproduct.z / crossproduct.length())

- use a more standard lighting model (e.g. diffuse/specular/phong)
  as described here: https://learnopengl.com/Lighting/Basic-Lighting

- let the user provide a function (normal vector -> lighting value) in
  the showShadow option

Note that shadow for surface plots was probably not used by many users,
because before this PR this did require manually unsetting
graph.surfaceColors after graph initialization.